requirements.txt
airsim==1.8.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
importlib-resources==6.4.0
intel-openmp==2021.4.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl==2021.4.0
mpmath==1.3.0
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.13.1
tbb==2021.13.0
threadpoolctl==3.5.0
torch==2.3.1
torchvision==0.18.1
tornado==4.5.3
typing-extensions==4.12.2
zipp==3.19.2
